revisi

+++ resources/views/admin/lowongan_kerja/components/actions.blade.php
@@ -1,4 +1,9 @@
 <div class="btn-group">
-    <button class="btn btn-sm btn-primary" onclick="editCustomer({{ $loker->id }})">Edit</button>
-    <button class="btn btn-sm btn-danger " onclick="deleteCustomers({{ $loker->id }})">Delete</button>
+    @if (Auth::user()->role == 'Perusahaan')
+        <a href="{{ route('lowongan-kerja.detail', $loker->id) }}" class="btn btn-sm btn-success ">Detail</a>
+        <button class="btn btn-sm btn-primary" onclick="editCustomer({{ $loker->id }})">Edit</button>
+        <button class="btn btn-sm btn-danger " onclick="deleteCustomers({{ $loker->id }})">Delete</button>
+    @else
+        <a href="{{ route('lowongan-kerja.detail', $loker->id) }}" class="btn btn-sm btn-success ">Detail</a>
+    @endif
 </div>
